# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "UDNS WEB 后台服务，主要提供域名管理、线路管理、后台管理等接口功能。",
        "title": "UDNS-WEB api服务",
        "termsOfService": "http://test.udns.woa.com",
        "contact": {},
        "version": "1.0"
    },
    "host": "test.udns.woa.com",
    "basePath": "/api/v1/",
    "paths": {
...
        "/zone": {
            "get": {
                "description": "根据名称匹配查询二级zone列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "二级zone管理"
                ],
                "summary": "二级zone列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "StaffName",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "page size",
                        "name": "ps",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "page number",
                        "name": "pn",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "基础域关键字，支持模糊匹配，可为空",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/frame.AuthResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "传递参数新增二级zone",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "二级zone管理"
                ],
                "summary": "新增二级zone",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "StaffName",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "新增二级zone参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/zone.Zone"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/frame.AuthResponse"
                        }
                    }
                }
            }
        },
        "/zone/{id}": {
            "put": {
                "description": "根据ID编辑具体二级zone",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "二级zone管理"
                ],
                "summary": "编辑二级zone",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "StaffName",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "编辑二级zone参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/zone.Zone"
                        }
                    },
                    {
                        "type": "string",
                        "description": "二级zone ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/frame.AuthResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "根据ID删除具体二级zone",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "二级zone管理"
                ],
                "summary": "删除二级zone",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "StaffName",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "二级zone ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/frame.AuthResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "frame.AuthResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "zone.Zone": {
            "type": "object",
            "properties": {
                "created_time": {
                    "description": "触发时间",
                    "type": "integer"
                },
                "deleted": {
                    "description": "是否删除",
                    "type": "boolean"
                },
                "deleted_time": {
                    "description": "删除时间",
                    "type": "integer"
                },
                "description": {
                    "description": "描述信息",
                    "type": "string"
                },
                "enabled": {
                    "description": "是否启用",
                    "type": "boolean"
                },
                "id": {
                    "description": "ID",
                    "type": "string"
                },
                "name": {
                    "description": "二级zone域名",
                    "type": "string"
                },
                "operator": {
                    "description": "负责人",
                    "type": "string"
                },
                "updated_time": {
                    "description": "恢复时间",
                    "type": "integer"
                },
                "zone_id": {
                    "description": "二级zone id",
                    "type": "integer"
                }
            }
        }
    }
}